Split tests into multiple feature files
Add tests common
Fix formatting
